[MCKIN-8436] Make API clients more generic
